move tests to servlet
