+++ src/test/java/com/google/error_analyzer/servlets/DocCountServletTest.java
@@ -8,7 +8,7 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.*/
-package com.google.error_analyzer;
+package com.google.error_analyzer.servlets;
 
 import com.google.error_analyzer.backend.LogDao;
 import com.google.error_analyzer.data.constant.LogFields;
